rename use-tag-selector-filter
